Force HomeKit update when values change upstream
While upstream Dreo app setting updates are reflected in the internal
state of `FanAccessory`, this is not enough to trigger an update in
HomeKit, which happen later, such as when refreshing the app, or when
reloading the Accessories page in HomeBridge UI.

This change explicitly calls
[updateValue](https://developers.homebridge.io/#/api/characteristics#characteristicupdatevalue)
on the corresponding characteristic in response to this external
trigger, which causes the accessory state to update instantaneously.

Tested this with DR-HAF003S on a local Homebridge installation.